// src/types/partials/SampledData.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TQuantity } from '../partials/Quantity';
import { TDecimal } from '../simpleTypes/Decimal';
import { TInt } from '../simpleTypes/Int';

export type TSampledData = {
    id?: String;
    extension?: TExtension[];
    origin: TQuantity;
    period: TDecimal;
    factor?: TDecimal;
    lowerLimit?: TDecimal;
    upperLimit?: TDecimal;
    dimensions: TInt;
    data?: String;
};